+++ src/monster.cpp
@@ -56,9 +56,9 @@ Monster::Monster(MonsterType* mType) :
 //	healthMax = mType->info.healthMax*multiplier;
 //	baseSpeed = mType->info.baseSpeed; 
 	
-    health = mType->info.health + (mType->info.health * (g_config.getDouble(ConfigManager::MONSTERLEVEL_BONUSHEALTH) * level));
-	healthMax = mType->info.healthMax + (mType->info.healthMax * (g_config.getDouble(ConfigManager::MONSTERLEVEL_BONUSHEALTH) * level));
-	baseSpeed = mType->info.baseSpeed + (mType->info.baseSpeed * (g_config.getDouble(ConfigManager::MONSTERLEVEL_BONUSSPEED) * level));	
+    health = mType->info.health + (mType->info.health * (g_config.getFloat(ConfigManager::MONSTERLEVEL_BONUSHEALTH) * level));
+	healthMax = mType->info.healthMax + (mType->info.healthMax * (g_config.getFloat(ConfigManager::MONSTERLEVEL_BONUSHEALTH) * level));
+	baseSpeed = mType->info.baseSpeed + (mType->info.baseSpeed * (g_config.getFloat(ConfigManager::MONSTERLEVEL_BONUSSPEED) * level));	
 	
 
 	internalLight = mType->info.light;
@@ -1922,7 +1922,7 @@ void Monster::dropLoot(Container* corpse, Creature*)
 {
 	if (corpse && lootDrop) {
 		//mType->createLoot(corpse);
-		mType->createLoot(corpse, g_config.getDouble(ConfigManager::MONSTERLEVEL_BONUSLOOT) * level);
+		mType->createLoot(corpse, g_config.getFloat(ConfigManager::MONSTERLEVEL_BONUSLOOT) * level);
 	}
 }
